requirements.txt
--index-url https://download.pytorch.org/whl/cu128

# ----------------------------
# Core utilities
# ----------------------------
python-dotenv
requests
aiohttp
websockets
psutil
pyyaml
tqdm
rich
pycaps[all] @ git+https://github.com/francozanardi/pycaps.git

# ----------------------------
# Numeric / ML foundations
# ----------------------------
numpy<2.0
scipy<1.14
einops
hdbscan

# ----------------------------
# PyTorch stack (CUDA 12.8 – pinned)
# ----------------------------
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128
torchsde

# ----------------------------
# Model / inference tooling
# ----------------------------
safetensors
huggingface-hub
transformers
accelerate
openai
openai-whisper
insightface

# ----------------------------
# Image / video processing
# ----------------------------
pillow
opencv-python-headless
imageio
imageio-ffmpeg
av
kornia

# ----------------------------
# ComfyUI internals / helpers
# ----------------------------
comfy_kitchen
omegaconf

# ----------------------------
# Google / YouTube APIs
# ----------------------------
google-api-python-client
google-auth
google-auth-oauthlib

# ----------------------------
# Server / API (used by ComfyUI)
# ----------------------------
fastapi
uvicorn
comfyui-frontend-package
